Extra notifications, data update/delete for user only or admin

+++ API/Controllers/AdminController.cs
@@ -82,11 +82,40 @@ public class AdminController : BaseApiController
     public async Task<ActionResult> DeleteUser(string username)
     {
         var user = await _unitOfWork.UsersRepository.GetUserByUsernameAsync(username);
+        if (user == null) return NotFound();
 
         var avatarPublicId = user.Avatar.PublicId;
         if (avatarPublicId != null) await _imageService.DeleteImageAsync(avatarPublicId);
 
         await _unitOfWork.MessagesRepository.DeleteUserMessagesAsync(username);
+        
+        var games = await _unitOfWork.GamesRepository.GetGamesForUserAsync(User.GetUserId());
+        foreach (var game in games)
+        {
+            var posterPublicId = game.Poster.PublicId;
+            if (posterPublicId != null)
+            {
+                var result = await _imageService.DeleteImageAsync(posterPublicId);
+                if (result.Error != null) return BadRequest(result.Error.Message);
+                game.Poster.Url = "";
+                game.Poster.PublicId = null;
+            }
+
+            foreach (var screenshot in game.Screenshots)
+            {
+                if (screenshot.PublicId != null)
+                {
+                    var result = await _imageService.DeleteImageAsync(screenshot.PublicId);
+                    if (result.Error != null) return BadRequest(result.Error.Message);
+                }
+            }
+            game.Screenshots.Clear();
+
+            _unitOfWork.GamesRepository.DeleteGame(game);
+            
+            _notificationCenter.GameDeleted(User.GetUsername(), game.Id);
+        }
+
         await _unitOfWork.Complete();
 
         await _userManager.DeleteAsync(user);
@@ -114,6 +143,44 @@ public class AdminController : BaseApiController
     }
 
     [Authorize(Policy = "AdminModeratorRole")]
+    [HttpDelete("delete-game/{title}")]
+    public async Task<ActionResult> DeleteGame(string title)
+    {
+        var game = await _unitOfWork.GamesRepository.GetGameByTitleAsync(title);
+        if (game == null) return NotFound();
+
+        var posterPublicId = game.Poster.PublicId;
+        if (posterPublicId != null)
+        {
+            var deletingResult = await _imageService.DeleteImageAsync(posterPublicId);
+            if (deletingResult.Error != null) return BadRequest(deletingResult.Error.Message);
+            game.Poster.Url = "";
+            game.Poster.PublicId = null;
+        }
+
+        foreach (var screenshot in game.Screenshots)
+        {
+            if (screenshot.PublicId != null)
+            {
+                var result = await _imageService.DeleteImageAsync(screenshot.PublicId);
+                if (result.Error != null) return BadRequest(result.Error.Message);
+            }
+        }
+        game.Screenshots.Clear();
+
+        _unitOfWork.GamesRepository.DeleteGame(game);
+
+        if (await _unitOfWork.Complete())
+        {
+            _notificationCenter.GameDeleted(User.GetUsername(), game.Id);
+
+            return Ok();
+        }
+
+        return BadRequest("Failed to delete game");
+    }
+
+    [Authorize(Policy = "AdminModeratorRole")]
     [HttpGet("reviews-for-moderation")]
     public async Task<ActionResult<PagedList<ReviewModerationDto>>> GetReviewsForModeration(
         [FromQuery]PaginationParams paginationParams)
@@ -134,14 +201,37 @@ public class AdminController : BaseApiController
     [HttpPut("approve-review/{id}")]
     public async Task<ActionResult> ApproveReview(int id)
     {
-        var review = await _unitOfWork.ReviewsRepository.ApproveReview(id);
+        var review = await _unitOfWork.ReviewsRepository.GetReviewByIdAsync(id);
+        if (review == null) return NotFound();
+
+        review.IsApproved = true;
 
         if (await _unitOfWork.Complete())
         {
-            _notificationCenter.ReviewPosted(User.GetUsername(), review);
+            _notificationCenter.ReviewPosted(User.GetUsername(), _mapper.Map<ReviewDto>(review));
+
             return Ok();
         }
 
         return BadRequest("Failed to approve review");
+    }
+
+    [Authorize(Policy = "AdminModeratorRole")]
+    [HttpDelete("reject-review/{id}")]
+    public async Task<ActionResult> RejectReview(int id)
+    {
+        var review = await _unitOfWork.ReviewsRepository.GetReviewByIdAsync(id);
+        if (review == null) return NotFound();
+
+        _unitOfWork.ReviewsRepository.DeleteReview(review);
+
+        if (await _unitOfWork.Complete())
+        {
+            _notificationCenter.ReviewDeleted(User.GetUsername(), id);
+            
+            return Ok();
+        }
+        
+        return BadRequest("Failed to delete review");
     }
 }
